% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen-download-files.R
\name{zen_download_files}
\alias{zen_download_files}
\title{Download files of a Zenodo repository}
\usage{
zen_download_files(record_id, token, path = ".", files = NULL)
}
\arguments{
\item{record_id}{a \code{character} of length 1. The Zenodo identifier of the
repository (7 numbers). Can be found on the URL of the repository}

\item{token}{a \code{character} of length 1. A token generated by Zenodo to
access files of a restricted access.}

\item{path}{a \code{character} of length 1. The folder in which files will be
saved.}

\item{files}{a \code{character} vector. If one or several files are provided,
only these files will be downloaded. If \code{NULL} (default) all files will be
downloaded.}
}
\value{
No returned value.
}
\description{
Downloads files of a Zenodo repository. If the repository is in restricted
access, a token (argument \code{token}) must be provided.
}
\examples{
# List files available in the repo: https://zenodo.org/record/7936568 ----
zen_list_files("7936568")

\dontrun{
# Download one file ----
zen_download_files("7936568", files = "FORCIS_pump_v03_14062023.csv")

# Download all files ----
zen_download_files("7936568")
}
}
