% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen-list-files.R
\name{zen_list_files}
\alias{zen_list_files}
\title{List files of a Zenodo repository}
\usage{
zen_list_files(record_id, token)
}
\arguments{
\item{record_id}{a \code{character} of length 1. The Zenodo identifier of the
repository (7 numbers). Can be found on the URL of the repository}

\item{token}{a \code{character} of length 1. A token generated by Zenodo to
access files of a restricted access.}
}
\value{
No returned value.
}
\description{
Lists files of a Zenodo repository. If the repository is in restricted
access, a token (argument \code{token}) must be provided.
}
\examples{
# List files available in the repo: https://zenodo.org/record/7936568 ----
zen_list_files("7936568")
}
